+++ addenda02_test.go
@@ -1,0 +1,31 @@
+// Copyright 2018 The ACH Authors
+// Use of this source code is governed by an Apache License
+// license that can be found in the LICENSE file.
+
+package ach
+
+import (
+	"testing"
+)
+
+func mockAddenda02() *Addenda02 {
+	addenda02 := NewAddenda02()
+	addenda02.ReferenceInformationOne = "REFONEA"
+	addenda02.ReferenceInformationTwo = "REF"
+	addenda02.TerminalIdentificationCode = "TERM02"
+	addenda02.TransactionSerialNumber = "100049"
+	addenda02.TransactionDate = "0612"
+	addenda02.AuthorizationCodeOrExpireDate = "123456"
+	addenda02.TerminalLocation = "Target Store 0049"
+	addenda02.TerminalCity = "PHILADELPHIA"
+	addenda02.TerminalState = "PA"
+	addenda02.TraceNumber = 91012980000088
+	return addenda02
+}
+
+func TestMockAddenda02(t *testing.T) {
+	addenda02 := mockAddenda02()
+	if err := addenda02.Validate(); err != nil {
+		t.Error("mockAddenda02 does not validate and will break other tests")
+	}
+}
