+++ batchPOS_test.go
@@ -33,21 +33,6 @@ func mockPOSEntryDetail() *EntryDetail {
 	return entry
 }
 
-func mockAddenda02() *Addenda02 {
-	addenda02 := NewAddenda02()
-	addenda02.ReferenceInformationOne = "REFONEA"
-	addenda02.ReferenceInformationTwo = "REF"
-	addenda02.TerminalIdentificationCode = "TERM02"
-	addenda02.TransactionSerialNumber = "100049"
-	addenda02.TransactionDate = "0612"
-	addenda02.AuthorizationCodeOrExpireDate = "123456"
-	addenda02.TerminalLocation = "Target Store 0049"
-	addenda02.TerminalCity = "PHILADELPHIA"
-	addenda02.TerminalState = "PA"
-	addenda02.TraceNumber = 91012980000088
-	return addenda02
-}
-
 // mockBatchPOS creates a BatchPOS
 func mockBatchPOS() *BatchPOS {
 	mockBatch := NewBatchPOS(mockBatchPOSHeader())
